Revert "Update globals.h"
This reverts commit 290922d6e48e589aaf7c76be2eda1021364a2e1a.

+++ src/globals.h
@@ -98,7 +98,7 @@ extern "C" {
 //    these are defined in errno.h
 #define ERR_BASE_RNC      -90
 
-unsigned long long LbSystemClockMilliseconds(void);
+uint64_t LbSystemClockMilliseconds(void);
 
 // Debug fuction-like macros - for free messages
 #define ERRORMSG(format, ...) LbErrorLog(format "\n", ##__VA_ARGS__)
